Refactoring for ResultsView. Move code into a few separate components.
