+++ manoseimas/compatibility_test/client/components/index.js
@@ -1,12 +1,12 @@
-import StatusBar from './StatusBar'
 import Block from './Block'
 import Button from './Button'
 import ButtonsBlock from './ButtonsBlock'
 import Checkbox from './Checkbox'
+import SimilarityBar from './SimilarityBar'
+import StatusBar from './StatusBar'
 import StickyHeader from './StickyHeader'
 import Tabs from './Tabs'
 import TopicGroup from './TopicGroup'
-import SimilarityBar from './SimilarityBar'
 
 export {
     Block,
